ref: Restructure code a bit more
